+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 9.2.0 / 2020-01-21
 * fix PHP warning [#823](https://github.com/afragen/github-updater/issues/823), thanks @pbiron
 * remove scheduled cron events on deactivation
 * added function to rename or recursively copy from `$source` to `$destination` and remove files/directories after copying. Should be more versatile than `$wp_filesystem->move()`. Fixes [#826](https://github.com/afragen/github-updater/issues/826)